Altered level 2 and 3 sizes
The level 2 and 3 had a scale of two which doesn't match the others. I tried to fix them. There is one more level that also has a scale of 2 on the tilemap but I should have probably asked before doing these two so I'll leave it as is for now